Add header and move content to new component
