+++ src/App.js
@@ -73,6 +73,7 @@ return (
           <Errors states={propStates} />
           <Switch>
             <Route exact path={["/blocks", "/"]} render={(props) => <MainView states={{ ...propStates, ...props }} />} />
+            <Route exact path="/about" render={(props) => <MainView states={{ ...propStates, ...props }} />} />
             <Route exact path="/block" render={(props) => <MainView states={{ ...propStates, ...props }} />} />
             <Route exact path="/tx" render={(props) => <MainView states={{ ...propStates, ...props }} />} />
             <Route exact path="/data" render={(props) => <MainView states={{ ...propStates, ...props }} />} />
